requirements.txt
asgiref==3.4.1
Django==3.2.7
django-dotenv==1.4.2
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.3
pytz==2021.1
sqlparse==0.4.2
